Experiment with user models
Experimenting with user domain objects. I'm not yet sure if I
like this. Historically, I've avoided this sort of thing in
client-side JS apps, but it's kinda handy with typescript. OTOH,
there are some advantages to just passing dumb resource object
to simple, composable functions.

Perhaps it'd be best to adopt a hybrid approach, specifying the
properties I intend to use with interfaces but moving the behavior
into stand-alone functions (or maybe a service object).